psql rust cache test refactor
